pipeline_automation/requirements.txt
pandas
pandas-gbq
google-cloud-bigquery
gspread
oauth2client
db-dtypes